chore(sdk-coin-xlm): move bitgo code
- moves xlm & stellarToken to xlm coin module

Ticket: BG-51039
