+++ modules/bitgo/src/v2/coins/index.ts
@@ -42,6 +42,8 @@ import { Stx, Tstx } from '@bitgo/sdk-coin-stx';
 export { Stx, Tstx };
 import { Trx, Ttrx } from '@bitgo/sdk-coin-trx';
 export { Trx, Ttrx };
+import { StellarToken, Txlm, Xlm } from '@bitgo/sdk-coin-xlm';
+export { StellarToken, Txlm, Xlm };
 import { Xrp, Txrp } from '@bitgo/sdk-coin-xrp';
 export { Xrp, Txrp };
 import { Xtz, Txtz } from '@bitgo/sdk-coin-xtz';
@@ -52,7 +54,6 @@ export { Zec, Tzec };
 export { Eth2 } from './eth2';
 export { Ofc } from './ofc'; // TODO: SDKT-15: make abstract and remove from exported list
 export { OfcToken } from './ofcToken';
-export { StellarToken } from './stellarToken';
 export { Susd } from './susd';
 export { FiatUsd } from './fiatusd';
 export { FiatEur } from './fiateur';
@@ -62,7 +63,5 @@ export { TfiatUsd } from './tfiatusd';
 export { TfiatEur } from './tfiateur';
 export { Hbar } from './hbar';
 export { Thbar } from './thbar';
-export { Txlm } from './txlm';
-export { Xlm } from './xlm';
 export { Polygon } from './polygon';
 export { Tpolygon } from './tpolygon';
