+++ modules/sdk-coin-xlm/src/index.ts
@@ -1,2 +1,4 @@
-export * from './xlm';
+export * from './getStellarKeys';
+export * from './stellarToken';
 export * from './txlm';
+export * from './xlm';
